Refactor animation code to group viewport functionality together
