+++ src/animation/utils/viewport-pixels.ts
@@ -1,3 +1,19 @@
+// Support types for viewport centric data
+type ViewportUnit = 'vw' | 'vh' | 'vmin' | 'vmax';
+
+export type ViewportVar = {
+  value: number;
+  unit: ViewportUnit;
+  varName: string;
+};
+
+export type RawViewportVar = [number, ViewportUnit];
+export type RawViewportPosition = {
+  x: RawViewportVar[];
+  y: RawViewportVar[];
+};
+
+
 // Create div with viewport offsets to determine true pixel location (innerWidth / innerHeight can't be trusted on mobile Safari)
 const measurementProxy = document.createElement('div');
 measurementProxy.setAttribute('style', 'bottom: calc(100vh + 0px); right: calc(100vw + 0px)');
@@ -29,4 +45,34 @@ export const getViewportHeight: () => number = () => {
     cacheDimensions();
   }
   return height;
+};
+
+
+// Export convenience methods for converting between viewport and pixel dimensions
+export const viewportUnitToPx = (value: number, unit: ViewportUnit): number => {
+  if (value === 0) return 0;
+  let scale = 0;
+  switch (unit) {
+    case 'vw':
+      scale = getViewportWidth();
+      break;
+    case 'vh':
+      scale = getViewportHeight();
+      break;
+    case 'vmin':
+      scale = Math.min(getViewportWidth(), getViewportHeight());
+      break;
+    case 'vmax':
+      scale = Math.max(getViewportWidth(), getViewportHeight());
+      break;
+  }
+  return value * scale * 0.01;
+};
+
+export const viewportVarChainToPx = (viewportVars: ViewportVar[]): number => {
+  let tally = 0;
+  for (let viewVar of viewportVars) {
+    tally += viewportUnitToPx(viewVar.value, viewVar.unit);
+  }
+  return tally;
 };
